cleanup the root.go file
